+++ packages/core-js/features/symbol/index.js
@@ -3,4 +3,3 @@ module.exports = require('../../es/symbol');
 require('../../modules/esnext.symbol.description');
 require('../../modules/esnext.symbol.observable');
 require('../../modules/esnext.symbol.pattern-match');
-require('../../modules/esnext.symbol.pattern-value');