+++ packages/core-js/features/symbol/pattern-value.js
@@ -1,3 +1,0 @@
-require('../../modules/esnext.symbol.pattern-value');
-
-module.exports = require('../../internals/wrapped-well-known-symbol').f('patternValue');