+++ packages/core-js/modules/esnext.symbol.pattern-value.js
@@ -1,3 +1,0 @@
-// `Symbol.patternValue` well-known symbol
-// https://github.com/tc39/proposal-pattern-matching
-require('../internals/define-well-known-symbol')('patternValue');