+++ packages/core-js/proposals/pattern-matching.js
@@ -1,2 +1,1 @@
 require('../modules/esnext.symbol.pattern-match');
-require('../modules/esnext.symbol.pattern-value');