+++ tests/pure/index.js
@@ -190,7 +190,6 @@ import './esnext.string.trim-left';
 import './esnext.string.trim-right';
 import './esnext.symbol.observable';
 import './esnext.symbol.pattern-match';
-import './esnext.symbol.pattern-value';
 import './esnext.weak-map.from';
 import './esnext.weak-map.of';
 import './esnext.weak-set.from';